+++ percy/parser/recipe_parser.py
@@ -1,5 +1,14 @@
-def RecipeParser():
+"""
+File:           recipe_parser.py
+Description:    Legacy stubs for anaconda-linter compatibility
+"""
+
+
+def RecipeParser() -> None:
+    "Stub"
     pass
 
-def SelectorConflictMode():
+
+def SelectorConflictMode() -> None:
+    "Stub"
     pass