+++ percy/render/_renderer.py
@@ -90,11 +90,39 @@ class _JinjaSilentUndefined(jinja2.Undefined):
 
         return EmptyString()
 
-    __add__ = __radd__ = __mul__ = __rmul__ = __div__ = __rdiv__ = __truediv__ = __rtruediv__ = __floordiv__ = (
+    __add__ = (
+        __radd__
+    ) = (
+        __mul__
+    ) = (
+        __rmul__
+    ) = (
+        __div__
+    ) = (
+        __rdiv__
+    ) = (
+        __truediv__
+    ) = (
+        __rtruediv__
+    ) = (
+        __floordiv__
+    ) = (
         __rfloordiv__
-    ) = __mod__ = __rmod__ = __pos__ = __neg__ = __call__ = __getitem__ = __lt__ = __le__ = __gt__ = __ge__ = (
-        __int__
-    ) = __float__ = __complex__ = __pow__ = __rpow__ = _fail_with_undefined_error
+    ) = (
+        __mod__
+    ) = (
+        __rmod__
+    ) = (
+        __pos__
+    ) = (
+        __neg__
+    ) = (
+        __call__
+    ) = (
+        __getitem__
+    ) = (
+        __lt__
+    ) = __le__ = __gt__ = __ge__ = __int__ = __float__ = __complex__ = __pow__ = __rpow__ = _fail_with_undefined_error
 
 
 _jinja_silent_undef = jinja2.Environment(undefined=_JinjaSilentUndefined)
@@ -247,7 +275,7 @@ def render(
             yaml_jinja.allow_duplicate_keys = True
             yaml_jinja.width = 1000
             data = None
-            with open(recipe_file) as fp:
+            with open(recipe_file, encoding="utf-8") as fp:
                 data = yaml_jinja.load(fp)
             return data
         else: